rsa and forms out of main
