+++ src/blockchain/mod.rs
@@ -10,7 +10,7 @@ use crate::blockchain::OperationCode::{
     Accept, Endorse, Issue, Mint, RequestToAccept, RequestToPay, Sell,
 };
 use crate::constants::BILLS_FOLDER_PATH;
-use crate::{encrypt_bytes, BitcreditBill, IdentityPublicData};
+use crate::{util::rsa::encrypt_bytes, BitcreditBill, IdentityPublicData};
 
 pub use block::Block;
 pub use chain::Chain;
