+++ src/util/mod.rs
@@ -1,1 +1,16 @@
+use std::fs::DirEntry;
+
 pub mod numbers_to_words;
+pub mod rsa;
+
+pub unsafe fn structure_as_u8_slice<T: Sized>(p: &T) -> &[u8] {
+    std::slice::from_raw_parts((p as *const T) as *const u8, size_of::<T>())
+}
+
+pub fn is_not_hidden(entry: &DirEntry) -> bool {
+    entry
+        .file_name()
+        .to_str()
+        .map(|s| !s.starts_with("."))
+        .unwrap_or(false)
+}
