+++ src/util/rsa.rs
@@ -1,0 +1,227 @@
+use openssl::pkey::{Private, Public};
+use openssl::rsa;
+use openssl::rsa::{Padding, Rsa};
+
+pub fn generation_rsa_key() -> Rsa<Private> {
+    Rsa::generate(2048).unwrap()
+}
+
+pub fn pem_private_key_from_rsa(rsa: &Rsa<Private>) -> String {
+    let private_key: Vec<u8> = rsa.private_key_to_pem().unwrap();
+    String::from_utf8(private_key).unwrap()
+}
+
+pub fn pem_public_key_from_rsa(rsa: &Rsa<Private>) -> String {
+    let public_key: Vec<u8> = rsa.public_key_to_pem().unwrap();
+    String::from_utf8(public_key).unwrap()
+}
+
+pub fn private_key_from_pem_u8(private_key_u8: &[u8]) -> Rsa<Private> {
+    rsa::Rsa::private_key_from_pem(private_key_u8).unwrap()
+}
+
+pub fn public_key_from_pem_u8(public_key_u8: &[u8]) -> Rsa<Public> {
+    rsa::Rsa::public_key_from_pem(public_key_u8).unwrap()
+}
+//--------------------------------------------------------------
+
+//-------------------------Bytes common-------------------------
+pub fn encrypt_bytes_with_public_key(bytes: &[u8], public_key: String) -> Vec<u8> {
+    let public_key = Rsa::public_key_from_pem(public_key.as_bytes()).unwrap();
+
+    let key_size: usize = (public_key.size() / 2) as usize; //128
+
+    let mut whole_encrypted_buff: Vec<u8> = Vec::new();
+    let mut temp_buff: Vec<u8> = vec![0; key_size];
+    let mut temp_buff_encrypted: Vec<u8> = vec![0; public_key.size() as usize];
+
+    let number_of_key_size_in_whole_bill: usize = bytes.len() / key_size;
+    let remainder: usize = bytes.len() - key_size * number_of_key_size_in_whole_bill;
+
+    for i in 0..number_of_key_size_in_whole_bill {
+        for j in 0..key_size {
+            let byte_number: usize = key_size * i + j;
+            temp_buff[j] = bytes[byte_number];
+        }
+
+        let _encrypted_len: usize = public_key
+            .public_encrypt(&temp_buff, &mut temp_buff_encrypted, Padding::PKCS1)
+            .unwrap();
+
+        whole_encrypted_buff.append(&mut temp_buff_encrypted);
+        temp_buff = vec![0; key_size];
+        temp_buff_encrypted = vec![0; public_key.size() as usize];
+    }
+
+    if remainder != 0 {
+        temp_buff = vec![0; remainder];
+
+        let position: usize = key_size * number_of_key_size_in_whole_bill;
+        let mut index_in_temp_buff: usize = 0;
+
+        for i in position..bytes.len() {
+            temp_buff[index_in_temp_buff] = bytes[i];
+            index_in_temp_buff += 1;
+        }
+
+        index_in_temp_buff = 0;
+
+        let _encrypted_len: usize = public_key
+            .public_encrypt(&temp_buff, &mut temp_buff_encrypted, Padding::PKCS1)
+            .unwrap();
+
+        whole_encrypted_buff.append(&mut temp_buff_encrypted);
+        temp_buff.clear();
+        temp_buff_encrypted.clear();
+    }
+
+    whole_encrypted_buff
+}
+
+pub fn decrypt_bytes_with_private_key(bytes: &[u8], private_key: String) -> Vec<u8> {
+    let private_key = Rsa::private_key_from_pem(private_key.as_bytes()).unwrap();
+
+    let key_size: usize = private_key.size() as usize; //256
+
+    let mut whole_decrypted_buff: Vec<u8> = Vec::new();
+    let mut temp_buff: Vec<u8> = vec![0; private_key.size() as usize];
+    let mut temp_buff_decrypted: Vec<u8> = vec![0; private_key.size() as usize];
+
+    let number_of_key_size_in_whole_bill: usize = bytes.len() / key_size;
+    // let remainder = bill_bytes.len() - key_size * number_of_key_size_in_whole_bill;
+
+    for i in 0..number_of_key_size_in_whole_bill {
+        for j in 0..key_size {
+            let byte_number = key_size * i + j;
+            temp_buff[j] = bytes[byte_number];
+        }
+
+        let decrypted_len: usize = private_key
+            .private_decrypt(&temp_buff, &mut temp_buff_decrypted, Padding::PKCS1)
+            .unwrap();
+
+        whole_decrypted_buff.append(&mut temp_buff_decrypted[0..decrypted_len].to_vec());
+        temp_buff = vec![0; private_key.size() as usize];
+        temp_buff_decrypted = vec![0; private_key.size() as usize];
+    }
+
+    // if remainder != 0 {
+    //     let position = key_size * number_of_key_size_in_whole_bill;
+    //     let mut index_in_temp_buff = 0;
+    //
+    //     for i in position..bill_bytes.len() {
+    //         temp_buff[index_in_temp_buff] = bill_bytes[i];
+    //         index_in_temp_buff = index_in_temp_buff + 1;
+    //     }
+    //
+    //     index_in_temp_buff = 0;
+    //
+    //     let decrypted_len = rsa_key
+    //         .public_decrypt(&*temp_buff, &mut temp_buff_decrypted, Padding::PKCS1)
+    //         .unwrap();
+    //
+    //     whole_decrypted_buff.append(&mut temp_buff_decrypted);
+    //     temp_buff.clear();
+    //     temp_buff_decrypted.clear();
+    // }
+
+    whole_decrypted_buff
+}
+
+pub fn encrypt_bytes(bytes: &[u8], rsa_key: &Rsa<Private>) -> Vec<u8> {
+    let key_size: usize = (rsa_key.size() / 2) as usize; //128
+
+    let mut whole_encrypted_buff: Vec<u8> = Vec::new();
+    let mut temp_buff: Vec<u8> = vec![0; key_size];
+    let mut temp_buff_encrypted: Vec<u8> = vec![0; rsa_key.size() as usize];
+
+    let number_of_key_size_in_whole_bill: usize = bytes.len() / key_size;
+    let remainder: usize = bytes.len() - key_size * number_of_key_size_in_whole_bill;
+
+    for i in 0..number_of_key_size_in_whole_bill {
+        for j in 0..key_size {
+            let byte_number: usize = key_size * i + j;
+            temp_buff[j] = bytes[byte_number];
+        }
+
+        let _encrypted_len: usize = rsa_key
+            .public_encrypt(&temp_buff, &mut temp_buff_encrypted, Padding::PKCS1)
+            .unwrap();
+
+        whole_encrypted_buff.append(&mut temp_buff_encrypted);
+        temp_buff = vec![0; key_size];
+        temp_buff_encrypted = vec![0; rsa_key.size() as usize];
+    }
+
+    if remainder != 0 {
+        temp_buff = vec![0; remainder];
+
+        let position: usize = key_size * number_of_key_size_in_whole_bill;
+        let mut index_in_temp_buff: usize = 0;
+
+        for i in position..bytes.len() {
+            temp_buff[index_in_temp_buff] = bytes[i];
+            index_in_temp_buff += 1;
+        }
+
+        index_in_temp_buff = 0;
+
+        let _encrypted_len: usize = rsa_key
+            .public_encrypt(&temp_buff, &mut temp_buff_encrypted, Padding::PKCS1)
+            .unwrap();
+
+        whole_encrypted_buff.append(&mut temp_buff_encrypted);
+        temp_buff.clear();
+        temp_buff_encrypted.clear();
+    }
+
+    whole_encrypted_buff
+}
+
+pub fn decrypt_bytes(bytes: &[u8], rsa_key: &Rsa<Private>) -> Vec<u8> {
+    let key_size: usize = rsa_key.size() as usize; //256
+
+    let mut whole_decrypted_buff: Vec<u8> = Vec::new();
+    let mut temp_buff: Vec<u8> = vec![0; rsa_key.size() as usize];
+    let mut temp_buff_decrypted: Vec<u8> = vec![0; rsa_key.size() as usize];
+
+    let number_of_key_size_in_whole_bill: usize = bytes.len() / key_size;
+    // let remainder = bill_bytes.len() - key_size * number_of_key_size_in_whole_bill;
+
+    for i in 0..number_of_key_size_in_whole_bill {
+        for j in 0..key_size {
+            let byte_number = key_size * i + j;
+            temp_buff[j] = bytes[byte_number];
+        }
+
+        let decrypted_len: usize = rsa_key
+            .private_decrypt(&temp_buff, &mut temp_buff_decrypted, Padding::PKCS1)
+            .unwrap();
+
+        whole_decrypted_buff.append(&mut temp_buff_decrypted[0..decrypted_len].to_vec());
+        temp_buff = vec![0; rsa_key.size() as usize];
+        temp_buff_decrypted = vec![0; rsa_key.size() as usize];
+    }
+
+    // if remainder != 0 {
+    //     let position = key_size * number_of_key_size_in_whole_bill;
+    //     let mut index_in_temp_buff = 0;
+    //
+    //     for i in position..bill_bytes.len() {
+    //         temp_buff[index_in_temp_buff] = bill_bytes[i];
+    //         index_in_temp_buff = index_in_temp_buff + 1;
+    //     }
+    //
+    //     index_in_temp_buff = 0;
+    //
+    //     let decrypted_len = rsa_key
+    //         .public_decrypt(&*temp_buff, &mut temp_buff_decrypted, Padding::PKCS1)
+    //         .unwrap();
+    //
+    //     whole_decrypted_buff.append(&mut temp_buff_decrypted);
+    //     temp_buff.clear();
+    //     temp_buff_decrypted.clear();
+    // }
+
+    whole_decrypted_buff
+}
